requirements.txt
cffi==1.15.0
msgpack==1.0.2
py-algorand-sdk==1.8.0
pycparser==2.20
pycryptodomex==3.11.0
PyNaCl==1.4.0
python-dotenv==0.19.1
six==1.16.0
m2r2==0.3.1
numpy
pandas
requests
multiprocess
boto3
pyteal